feat: improve language activity display and strengthen GitHub synchronization resilience
Summary of changes:\n- Fixed language formatting and capitalization (e.g., TypeScript, C#).\n- Lowered language threshold to 10KB to ensure better visibility.\n- Improved frontend reactivity and data fetching on direct navigation.\n- Hardened backend GitHub API service to handle unexpected responses gracefully.\n- Implemented stale repository cleanup logic to ensure accurate data.\n- Fixed repository fetching endpoint to correctly target repo lists.

+++ backend/lib/models/database_models/GithubModel.py
@@ -41,7 +41,7 @@ class LanguageAssosiationModel(BASE):
 
     id = Column(Integer, primary_key = True, index = True, nullable = False, autoincrement = True)
     lang_id = Column(Integer, ForeignKey('languages.id'), nullable = False)
-    repo_id = Column(Integer, ForeignKey('repositories.repo_id'), nullable = False)
+    repo_id = Column(BigInteger, ForeignKey('repositories.repo_id'), nullable = False)
     code_bytes = Column(Integer, nullable = False)
 
     language = relationship("LanguageModel", back_populates = "assosiations")
